Fix getDutyCycle return type to make it possible to increase resolution

+++ flix/motors.ino
@@ -26,7 +26,7 @@ void setupMotors() {
 	Serial.println("Motors initialized");
 }
 
-uint8_t getDutyCycle(float value) {
+int getDutyCycle(float value) {
 	value = constrain(value, 0, 1);
 	float duty = mapff(value, 0, 1, 0, (1 << PWM_RESOLUTION) - 1);
 	return round(duty);
